Create guess_the_number_computer.py
In this code, the computer is not told the number. The range can be set as per user wishes and can decide any number in between. The computer will go on guessing a random number. All we have to do is suggest it whether it's current guess is higher or lower than the number they have to guess. It will again guess a number between the new range. This goes on until the computer guesses the number.